requirements.txt
Django==4.2.4
django-ninja==0.22.2
orjson==3.9.2
pytest==7.4.0
pytest-django==4.5.2
requests==2.31.0
pytest-factoryboy==2.5.1
Faker==19.3.0
django-ninja-jwt==5.2.5
django-ninja-extra==0.19.1
